backend/requirements.txt
celery==5.5.3
Flask==3.1.2
Flask_Caching==2.3.1
Flask_Cors==4.0.0
Flask_JWT_Extended==4.7.1
flask_mail==0.10.0
Flask_RESTful==0.3.10
flask_sqlalchemy==3.1.1
python-dotenv==1.1.1
Werkzeug==3.1.3
